horoscope_bot/requirements.txt
python-telegram-bot
requests
beautifulsoup4
lxml
python-dotenv
Pillow
